+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -128,8 +128,6 @@ export { BYOCWrapper };
 
 export { ComponentBuilder, ComponentBuilderConfig } from './ComponentBuilder';
 
-export { Context, ContextConfig, SDK } from './context';
-
 export {
   ComponentFactory,
   Image,