+++ packages/sitecore-jss-proxy/src/personalize/index.ts
@@ -1,0 +1,2 @@
+export { PersonalizeHelper } from './PersonalizeHelper';
+export { PersonalizeConfig } from '../types/personalize';