+++ packages/sitecore-jss-proxy/src/types/index.ts
@@ -1,2 +1,3 @@
 export { AppRenderer, RenderResponse } from './AppRenderer';
 export { RouteUrlParser } from './RouteUrlParser';
+export { PersonalizeConfig } from './personalize';